review and recipe update

+++ src/main/java/com/puccampinas/backendp5noname/dtos/ReviewDTO.java
@@ -1,4 +1,4 @@
 package com.puccampinas.backendp5noname.dtos;
 
-public record ReviewDTO(String title, String description, Integer rating, String recipeId) {
+public record ReviewDTO(String title, String description, Long rating, String recipeId) {
 }
